+++ blkshp_os/inventory/doctype/inventory_audit_storage_location/__init__.py
@@ -1,0 +1,5 @@
+"""
+Inventory Audit Storage Location child table package.
+"""
+
+