+++ blkshp_os/inventory/doctype/inventory_counting_task/__init__.py
@@ -1,0 +1,5 @@
+"""
+Inventory Counting Task child table package.
+"""
+
+