+++ blkshp_os/recipes/doctype/recipe/__init__.py
@@ -1,0 +1,5 @@
+"""
+Recipe DocType package.
+"""
+
+