+++ blkshp_os/recipes/doctype/recipe_ingredient/__init__.py
@@ -1,0 +1,5 @@
+"""
+Recipe Ingredient child table package.
+"""
+
+